ok, here is the deal:
the line with the mysql_fetch_object($res) was unnecessary.
i am not sure why it was there.  i replaced it with isset()
i mean i know why it was there, but dont know why isset()
was not used in the first place.  anyway, the result was that
it was stripping one record from the $res of every page because the
same $res was getting objected twice for every loop, every 20 loops.

mk?  bye.