OH KAY
let us see how this works out in terms of user authentication.  i removed the session variable for $back_to for no other reason than it was not always sticking.  i found that ALWAYS is a requisite for this business of authentication.

a rider:
at some point id started being used for article_id, im not sure when that was, but i like id for its short ness. so i added an article_id to id converter in the global as a temporary fix for any ambiguity.