much of this has to do with the bozo filter, or the multi.  i need to check it in so that i can work the kinks out with the database.
article2.phtml is an experament with the pear IT templating shits.  i
dunno if i give a rats ass yet.  it would be cool to have a text only
s.o.  but that IT stuff is a pain in the butt and not very fun.

added bozofilter to rollcall i think.  top.inc has some auth stuff in
it?  i forget.